backend/energy_monitor_api/requirements.txt
Django==4.2.7
djangorestframework==3.16.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.7.0
gunicorn==21.2.0
whitenoise==6.6.0
